Continuing the integer datatype conversion.
This version of the code successfully compiles, links, and runs on x86 Linux
with no options selected in 'mxconfig.h'.


git-svn-id: http://svn.csrri.iit.edu/mx/trunk@253 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_cryostream600_motor.c
@@ -699,7 +699,7 @@ mxd_cryostream600_motor_wait_for_response( MX_RECORD *rs232_record,
 		= "mxd_cryostream600_motor_wait_for_response()";
 
 	int i, max_retries;
-	unsigned long num_input_bytes_available;
+	uint32_t num_input_bytes_available;
 	unsigned long wait_ms;
 	mx_status_type mx_status;
