+++ libMx/i_ortec974.h
@@ -65,7 +65,7 @@ MX_API mx_status_type mxi_ortec974_write( MX_GENERIC *generic,
 					void *buffer, size_t count );
 MX_API mx_status_type mxi_ortec974_num_input_bytes_available(
 				MX_GENERIC *generic,
-				unsigned long *num_input_bytes_available );
+				uint32_t *num_input_bytes_available );
 MX_API mx_status_type mxi_ortec974_discard_unread_input(
 					MX_GENERIC *generic, int debug_flag);
 MX_API mx_status_type mxi_ortec974_discard_unwritten_output(
